requirements.txt
langgraph
fastapi
python-dotenv
langchain-openai
pydantic
routers
streamlit
requests
pydantic-settings
langchain
uvicorn
langsmith
pyowm
langchain-community
duckduckgo-search
faker
sqlalchemy
# psycopg2
langchain-pinecone
pinecone
langchain-core
rapidfuzz
psycopg2-binary